requirements.txt
SQLAlchemy~=2.0.40
aiogram~=3.19.0
asyncpg~=0.30.0
pytz~=2025.2
APScheduler~=3.11.0
python-dotenv~=1.1.0